pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aa-standingsrequests"
dynamic = ["version"]
description = "App for managing character standing requests, made for Alliance Auth"
readme = "README.md"
license = "GPL-3.0"
requires-python = ">=3.8"
authors = [
    { name = "Erik Kalkoken", email = "kaloken87@gmail.com" },
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "allianceauth>=3",
    "allianceauth-app-utils>=1.18",
    "django-eveuniverse>=0.19.1",
]

[project.urls]
Homepage = "https://gitlab.com/ErikKalkoken/aa-standingsrequests"
Source = "https://gitlab.com/ErikKalkoken/aa-standingsrequests"
Changelog = "https://gitlab.com/ErikKalkoken/aa-standingsrequests/-/blob/master/CHANGELOG.md"
Tracker = "https://gitlab.com/ErikKalkoken/aa-standingsrequests/-/issues"


[tool.hatch.version]
path = "standingsrequests/__init__.py"

[tool.hatch.build]
include = [
    "/standingsrequests",
]


[tool.isort]
profile = "black"
multi_line_output = 3
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "ALLIANCEAUTH",
    "FIRSTPARTY",
    "LOCALFOLDER"
]
known_allianceauth = ["allianceauth", "app_utils"]
known_django = ["django", "django_webtest", "esi", "eveuniverse"]

[tool.pylint.'MASTER']
ignore-patterns = ["test_.*.py,__init__.py", "generate_.*.py"]

[tool.pylint.'BASIC']
good-names = ["i", "j", "k", "x", "f", "ex"]

[tool.pylint.'FORMAT']
max-line-length = 100

[tool.pylint.'MESSAGES CONTROL']
disable = ["R", "C"]
